MAINT: refactored space sampling classes in separate module
